Complete Restructuring of document layout
Abandoning shelves in favor of a single book structure
Document is now framed as a "Guide" that starts out with a recap of the
value proposition and then goes into the introductory content including
first Rust and ODP concept tutorials.
It introduces "Tracks of ODP" as a jumping off spot / repository guide
in which documentation within the invididual repositories is linked (via
readme) and categorized into a matrix for easier discovery.
The content then goes into the practical exercises for EC and Patina
components, and toward an end-project goal of building the virtual
laptop like before.
Specifications are at the end (like an appendix).